requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Document processing
pdfplumber==0.10.3
python-docx==1.1.0
requests==2.31.0
python-multipart==0.0.6

# Vector database and embeddings
pinecone-client==2.2.4
openai==1.3.0

# HTTP client and utilities
aiohttp==3.9.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Hugging Face (if needed)
huggingface-hub==0.19.4

# Google APIs (if needed)
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
